+++ models/dao/OBAModelDAO.m
@@ -22,8 +22,9 @@
 #import "OBAModelDAOUserPreferencesImpl.h"
 #import "OBAPlacemark.h"
 #import "OBABookmarkGroup.h"
+#import "OBAAnalytics.h"
 
-const static int kMaxEntriesInMostRecentList = 10;
+const NSInteger kMaxEntriesInMostRecentList = 10;
 
 @interface OBAModelDAO ()
 
@@ -89,13 +90,7 @@ const static int kMaxEntriesInMostRecentList = 10;
 - (void) setOBARegion:(OBARegionV2*)newRegion {
     _region = newRegion;
     [_preferencesDao writeOBARegion:newRegion];
-    NSLog(@"Set Region: %@",newRegion.regionName);
-    [TestFlight passCheckpoint:[NSString stringWithFormat:@"Set Region: %@",newRegion.regionName]];
-    [[GAI sharedInstance].defaultTracker
-     send:[[GAIDictionaryBuilder createEventWithCategory:@"ui_action"
-                                                  action:@"set_region"
-                                                   label:[NSString stringWithFormat:@"Set Region: %@",newRegion.regionName]
-                                                   value:nil] build]];
+    [OBAAnalytics reportEventWithCategory:@"ui_action" action:@"set_region" label:[NSString stringWithFormat:@"Set Region: %@",newRegion.regionName] value:nil];
 }
 
 
@@ -158,7 +153,7 @@ const static int kMaxEntriesInMostRecentList = 10;
         
     }
     
-    NSUInteger over = [_mostRecentCustomApiUrls count] - kMaxEntriesInMostRecentList;
+    NSInteger over = [_mostRecentCustomApiUrls count] - kMaxEntriesInMostRecentList;
     for( int i=0; i<over; i++)
         [_mostRecentCustomApiUrls removeObjectAtIndex:([_mostRecentCustomApiUrls count]-1)];