+++ view_controllers/OBABookmarksViewController.m
@@ -21,6 +21,7 @@
 #import "UITableViewController+oba_Additions.h"
 #import "OBABookmarkGroup.h"
 #import "OBAEditBookmarkGroupViewController.h"
+#import "OBAAnalytics.h"
 
 @interface OBABookmarksViewController ()
 @property(strong) NSArray *bookmarks;
@@ -61,11 +62,8 @@
     // We reload the table here in case we are coming back from the user editing the label for a bookmark
     [self _refreshBookmarks];
     [self.tableView reloadData];
-    [TestFlight passCheckpoint:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker set:kGAIScreenName
-                                       value:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker
-     send:[[GAIDictionaryBuilder createAppView] build]];
+    
+    [OBAAnalytics reportScreenView:[NSString stringWithFormat:@"View: %@", [self class]]];
 }
 
 - (void)viewWillDisappear:(BOOL)animated {
@@ -106,6 +104,10 @@
     NSInteger total = self.bookmarks.count;
     total += [self numberOfRowsForBookmarkGroups];
     return MAX(total, 1);
+}
+
+- (BOOL)tableView:(UITableView *)tableView canEditRowAtIndexPath:(NSIndexPath *)indexPath {
+    return self.bookmarks.count > 0 || self.bookmarkGroups.count > 0;
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {