+++ view_controllers/OBASearchResultsListViewController.m
@@ -21,6 +21,7 @@
 #import "OBASearchResultsMapViewController.h"
 #import "OBAStopViewController.h"
 #import "UITableViewController+oba_Additions.h"
+#import "OBAAnalytics.h"
 
 
 @interface OBASearchResultsListViewController ()
@@ -58,12 +59,9 @@
 
 - (void)viewWillAppear:(BOOL)animated {
     [super viewWillAppear:animated];
-    [TestFlight passCheckpoint:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker set:kGAIScreenName
-                                       value:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker
-     send:[[GAIDictionaryBuilder createAppView] build]];
     [self reloadData];
+
+    [OBAAnalytics reportScreenView:[NSString stringWithFormat:@"View: %@", [self class]]];
 }
 
 - (void)viewWillDisappear:(BOOL)animated {