+++ view_controllers/OBAVehicleDetailsController.m
@@ -5,6 +5,7 @@
 #import "OBASituationsViewController.h"
 #import "OBALogger.h"
 #import "OBAPresentation.h"
+#import "OBAAnalytics.h"
 
 
 typedef enum {
@@ -49,11 +50,8 @@ typedef enum {
 
 - (void)viewWillAppear:(BOOL)animated {
     [super viewWillAppear:animated];
-    [TestFlight passCheckpoint:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker set:kGAIScreenName
-                                       value:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker
-     send:[[GAIDictionaryBuilder createAppView] build]];
+    
+    [OBAAnalytics reportScreenView:[NSString stringWithFormat:@"View: %@", [self class]]];
 }
 
 - (BOOL) isLoading {