+++ view_controllers/StopDetails/OBAEditStopBookmarkViewController.m
@@ -22,6 +22,7 @@
 #import "UITableViewController+oba_Additions.h"
 #import "OBABookmarkGroup.h"
 #import "OBAEditStopBookmarkGroupViewController.h"
+#import "OBAAnalytics.h"
 
 @implementation OBAEditStopBookmarkViewController
 
@@ -77,11 +78,7 @@
         [_requests addObject:request];
     }
 
-    [TestFlight passCheckpoint:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker set:kGAIScreenName
-                                       value:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker
-     send:[[GAIDictionaryBuilder createAppView] build]];
+    [OBAAnalytics reportScreenView:[NSString stringWithFormat:@"View: %@", [self class]]];
 }
 
 #pragma mark OBAModelServiceRequest