+++ view_controllers/StopDetails/OBAReportProblemViewController.m
@@ -2,6 +2,7 @@
 #import "OBAReportProblemWithStopViewController.h"
 #import "OBAReportProblemWithRecentTripsViewController.h"
 #import "UITableViewController+oba_Additions.h"
+#import "OBAAnalytics.h"
 
 @implementation OBAReportProblemViewController
 
@@ -31,11 +32,8 @@
 
 - (void)viewWillAppear:(BOOL)animated {
     [super viewWillAppear:animated];
-    [TestFlight passCheckpoint:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker set:kGAIScreenName
-                                       value:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker
-     send:[[GAIDictionaryBuilder createAppView] build]];
+    
+    [OBAAnalytics reportScreenView:[NSString stringWithFormat:@"View: %@", [self class]]];
 }
 
 #pragma mark -