+++ view_controllers/situations/OBASituationsViewController.m
@@ -10,6 +10,7 @@
 #import "OBASituationV2.h"
 #import "OBASituationViewController.h"
 #import "UITableViewController+oba_Additions.h"
+#import "OBAAnalytics.h"
 
 
 @implementation OBASituationsViewController
@@ -41,11 +42,8 @@
 
 - (void)viewWillAppear:(BOOL)animated {
     [super viewWillAppear:animated];
-    [TestFlight passCheckpoint:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker set:kGAIScreenName
-                                       value:[NSString stringWithFormat:@"View: %@", [self class]]];
-    [[GAI sharedInstance].defaultTracker
-     send:[[GAIDictionaryBuilder createAppView] build]];
+    
+    [OBAAnalytics reportScreenView:[NSString stringWithFormat:@"View: %@", [self class]]];
 }
 
 #pragma mark -